+++ src/services/karaManagement.ts
@@ -29,7 +29,7 @@ import { getState } from '../utils/state';
 import { checkMediaAndDownload } from './download';
 import { getKara, getKaras } from './kara';
 import { editKara } from './karaCreation';
-import { checkDownloadStatus, getRepo, getRepos } from './repo';
+import { getRepo, getRepos } from './repo';
 import { updateAllSmartPlaylists } from './smartPlaylist';
 import { getTag } from './tag';
 
@@ -37,7 +37,7 @@ const service = 'KaraManager';
 
 export async function createKaraInDB(kara: KaraFileV4, opts = { refresh: true }) {
 	const oldData = await insertKara(kara);
-	await Promise.all([updateKaraParents(kara.data), updateTags(kara.data), checkDownloadStatus([kara.data.kid])]);
+	await Promise.all([updateKaraParents(kara.data), updateTags(kara.data)]);
 	if (opts.refresh) {
 		if (!oldData.old_modified_at) {
 			await refreshKarasAfterDBChange('ADD', [kara.data]);