+++ src/services/repo.ts
@@ -8,7 +8,7 @@ import { baseChecksum, editKaraInStore, getStoreChecksum, sortKaraStore } from '
 import { updateDownloaded } from '../dao/download';
 import { deleteRepo, insertRepo, selectRepos, updateRepo } from '../dao/repo';
 import { getSettings, refreshAll, saveSetting } from '../lib/dao/database';
-import { initHooks, stopWatchingHooks } from '../lib/dao/hook';
+import { initHooks } from '../lib/dao/hook';
 import { refreshKaras } from '../lib/dao/kara';
 import { parseKara, writeKara } from '../lib/dao/karafile';
 import { readAllKaras } from '../lib/services/generation';
@@ -132,7 +132,8 @@ export async function updateAllRepos() {
 
 export async function checkDownloadStatus(kids?: string[]) {
 	profile('checkDownloadStatus');
-	logger.info(`Checking downloaded status of ${kids ? kids.length : 'all'} songs`, { service });
+	// Avoid spamming logs if we're only checking one song at a time
+	if (kids?.length > 1) logger.info(`Checking downloaded status of ${kids ? kids.length : 'all'} songs`, { service });
 	const karas = await getKaras({
 		q: kids ? `k:${kids.join(',')}` : undefined,
 		ignoreCollections: true,
@@ -154,7 +155,8 @@ export async function checkDownloadStatus(kids?: string[]) {
 	if (mediasExisting.length > 0) {
 		updateDownloaded(mediasExisting, 'DOWNLOADED');
 	}
-	logger.info('Finished checking downloaded status', { service });
+	// Avoid spamming logs if we're only checking one song at a time
+	if (kids?.length > 1) logger.info('Finished checking downloaded status', { service });
 	profile('checkDownloadStatus');
 }
 
@@ -666,7 +668,6 @@ export async function newGitRepo(repo: Repository) {
 	const baseDir = resolve(state.dataPath, repo.BaseDir);
 	const mediaDir = resolve(state.dataPath, repo.Path.Medias[0]);
 	if (pathIsContainedInAnother(baseDir, mediaDir)) throw 'Media folder is contained in base dir, move it first!';
-	await stopWatchingHooks();
 	await remove(baseDir);
 	await asyncCheckOrMkdir(baseDir);
 	const git = await setupGit(repo, false, true);